checked int types for kan_cnn

+++ src/models/wavKAN_CNN/KAN_CNN.jl
@@ -25,7 +25,7 @@ struct wavKAN_CNN
 end
 
 function KAN_CNN(in_channels::Int, out_channels::Int, encoder_wav_names, encoder_activations, decoder_wav_names, decoder_activations)
-    hidden_dim = parse(Int32, get(ENV, "hidden_dim", "40"))
+    hidden_dim = parse(Int64, get(ENV, "hidden_dim", "40"))
     norm = parse(Bool, get(ENV, "norm", "false"))
 
     encoder_list = [
